Fixed motors-on logging bug and changed SPI input to volatile. Other interrupt-based computations also need to be checked

+++ logging.c
@@ -141,11 +141,11 @@ void ProcessLogging(void)
     return;
   }
 
-  if (((int)FlightCtrlState & (int)FC_STATE_BIT_MOTORS_RUNNING)
-    && !logging_active_) OpenLogFile(0);
+  if ((FlightCtrlState() & FC_STATE_BIT_MOTORS_RUNNING) && !logging_active_)
+    OpenLogFile(0);
 
-  if (!((int)FlightCtrlState & (int)FC_STATE_BIT_MOTORS_RUNNING)
-    && logging_active_) CloseLogFile();
+  if (!(FlightCtrlState() & FC_STATE_BIT_MOTORS_RUNNING) && logging_active_)
+    CloseLogFile();
 
   // Open the file if logging is supposed to be active but the file is closed.
   if (logging_active_ && !file_.fs)
